+++ 2024/day_13/day13.cabal
@@ -1,0 +1,19 @@
+cabal-version: 3.0
+name: day13
+version: 0.1.0.0
+synopsis: Advent of Code 2024 Day 13
+description: Solution for Advent of Code 2024 Day 13
+license: BSD-3-Clause
+author: Your Name
+maintainer: your.email@example.com
+category: Development
+build-type: Simple
+
+executable day13
+    main-is: Solution.hs
+    build-depends: 
+        base >= 4.16 && < 5,
+        hmatrix,
+        parsec
+    default-language: Haskell2010
+    ghc-options: -Wall
